+++ src/main/java/com/jadebloom/goblin_api/shared/validation/ValidIdList.java
@@ -1,0 +1,36 @@
+package com.jadebloom.goblin_api.shared.validation;
+
+import static java.lang.annotation.ElementType.ANNOTATION_TYPE;
+import static java.lang.annotation.ElementType.FIELD;
+import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.ElementType.PARAMETER;
+import static java.lang.annotation.ElementType.TYPE_USE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import org.hibernate.validator.constraints.ConstraintComposition;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+import jakarta.validation.ReportAsSingleViolation;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+
+@NotNull
+@Size(min = 1)
+@ReportAsSingleViolation
+@Target({ FIELD, METHOD, PARAMETER, ANNOTATION_TYPE, TYPE_USE })
+@Retention(RUNTIME)
+@ConstraintComposition
+@Constraint(validatedBy = {})
+public @interface ValidIdList {
+
+	String message() default "The ID list should include at least one ID";
+
+	Class<?>[] groups() default {};
+
+	Class<? extends Payload>[] payload() default {};
+
+}