+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 - Added "delete all currencies" functionality.
 - Added "delete all expense categories" functionality.
 - Added "delete all expenses" functionality.
+- Added "delete multiple currencies" functionality.
+- Added "delete multiple expense categories" functionality.
+- Added "delete multiple expenses" functionality.
 ### Changed
 - Changed the payload and destination of update by ID functionality.
 - Added expense_category_name and currency_name to ExpenseDto schema.